Восстановление сессий в оперативную память выключено (для тестирования).
