+++ Seterator/Models/Session.cs
@@ -1,0 +1,15 @@
+using System;
+using System.ComponentModel.DataAnnotations;
+
+namespace Seterator.Models
+{
+    public class Session
+    {
+        [Key]
+        public string Id { get; set; }
+        public Guid? UserGuid { get; set; }
+        public bool Authenticated { get; set; }
+
+        public User User { get; set; }
+    }
+}
